+++ ClassCreatorPHP.php
@@ -132,6 +132,9 @@ class ClassCreatorPHP
         if (!file_exists($folder)) {
             mkdir($folder, 0777, true);
         }
+        if (!$this->settings->overwrite && file_exists($fullPath)) {
+            return;
+        }
         file_put_contents($fullPath, $content);
     }
